Move inner test class to top level
Reason is powermock 2.0.2 attempts to transform the inner classes
of the test each time the test is executed, resulting in javassist
throwing an IllegalStateException because transformed class was frozen.
